Safari-609.1.11.1/LayoutTests/editing/selection/range-surroundContents-with-preceding-node.html

<!DOCTYPE html>
<html>
  <body>
    <script src="../../resources/js-test-pre.js"></script>
    <script>
        window.onload = function(){
            var range = document.createRange();
            var rootNode = document.getElementById("root");
            range.setStart(rootNode, 4);
            range.setEnd(rootNode, 6);
            var parentElement = document.getElementById("targetForSurround");
            range.surroundContents(parentElement);

            shouldBe(range.startContainer.id, rootNode.id);
            shouldBe(range.startOffset.toString(), "3");
            shouldBe(range.endContainer.id, rootNode.id);
            shouldBe(range.startOffset.toString(), "3");
        };
    </script>
    <script src="../../resources/js-test-post.js"></script>
    <div id="root">
      <div id="targetForSurround"></div>
      <div id="description">Test that there is no crash when surroundContents is called with a node preceding the current selection.</div>
      <div id="trailingNode"></div>
    </div>
  </body>
</html>
